chore: Update .gitignore to ignore temporary and video files
This commit updates the .gitignore file to ignore temporary files that can be created if a process still has a handle open of a deleted file. It also adds the "*.mp4" pattern to ignore video files. This helps to keep the repository clean and prevents unnecessary files from being tracked.