refactor: Update video_transcode package to use video ID for transcoding
This commit refactors the video_transcode package to use the video ID instead of the video file path for transcoding. It updates the `ProcessMessage` function to retrieve the video object from the database using the ID, and passes the video object to the `TranscodeVideo` function. This change improves the reliability and maintainability of the transcoding process.