refactor: Add timestamp to watermarked video
This commit adds a timestamp to the watermarked video by updating the `saveWatermarkedVideo` function in the `addWatermark` module. The timestamp is set to the current time using the `time.Now()` function and is stored in the `ProcessedAt` field of the `video` object. This change ensures that the processed videos have a record of when they were watermarked, which can be useful for tracking and auditing purposes.