feat: Add final_messages package to send success message after saving watermarked video
This commit adds the final_messages package to the shared directory. It includes the SendSuccessMessage function, which is called after successfully saving a watermarked video. The function sends a success message to the SUCCESS_TOPIC using Kafka, providing information about the video. This change ensures that the system can communicate the successful completion of video watermarking to other parts of the application.