refactor: Update MySQL connection logic to support SQLite for testing
This commit modifies the MySQL connection logic in the `Connect()` function to support SQLite for testing purposes. It checks the value of the `ENV` environment variable and if it is set to "test", it opens a connection to an SQLite database instead of MySQL. This change allows for easier testing and isolation of the database layer.